requirements.txt
Django>=4.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
python-dateutil>=2.8.0
djangorestframework>=3.14.0

